requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
Pillow==10.4.0
html5lib==1.1
playwright==1.40.0